requirements.txt
django-allauth>=0.58.0
djangorestframework
requests
PyJWT
cryptography
django
strawberry-graphql[asgi]
strawberry-graphql-django
channels
gunicorn
uvicorn[standard]
whitenoise
psycopg2-binary
dj-database-url
python-decouple
django-filter
text-unidecode
channels
channels-redis
redis
daphne
cloudinary
httpx==1.2.3